Revert "move ownKeys"
This reverts commit 4263c44551ddc3eda8b75cd8a7986ff739981815.
