--- src/utils/index.js
@@ -1,1 +1,0 @@
-index.js
